LA/Baseline-RawNet2/requirements.txt
numba==0.48
numpy==1.17.0
librosa==0.7.2
pyyaml==5.3.1
tensorboardX==2.1
tqdm==4.64.1